refactor(lang): Decouple Frame and TypeSystem
