+++ pseudocode/lang/__init__.py
@@ -40,6 +40,8 @@ from typing import (
 from dataclasses import dataclass
 
 from .object import *
+from .types import *
+from .typesystem import *
 from . import (
     object, )
